@@ -36,20 +36,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "15.1.6",
-    "@angular/animations": "15.1.5",
-    "@angular/cdk": "15.1.5",
-    "@angular/common": "15.1.5",
-    "@angular/compiler": "15.1.5",
-    "@angular/core": "15.1.5",
+    "@angular/animations": "~15.2.4",
+    "@angular/cdk": "~15.2.3",
+    "@angular/common": "~15.2.4",
+    "@angular/compiler": "~15.2.4",
+    "@angular/core": "~15.2.4",
     "@angular/fire": "^7.5.0",
-    "@angular/forms": "15.1.5",
-    "@angular/google-maps": "^15.1.5",
-    "@angular/material": "15.1.5",
-    "@angular/platform-browser": "15.1.5",
-    "@angular/platform-browser-dynamic": "15.1.5",
-    "@angular/router": "15.1.5",
-    "@angular/service-worker": "15.1.5",
+    "@angular/forms": "~15.2.4",
+    "@angular/google-maps": "^15.2.3",
+    "@angular/material": "~15.2.3",
+    "@angular/platform-browser": "~15.2.4",
+    "@angular/platform-browser-dynamic": "~15.2.4",
+    "@angular/router": "~15.2.4",
+    "@angular/service-worker": "~15.2.4",
     "@fortawesome/angular-fontawesome": "^0.12.1",
     "@fortawesome/fontawesome-pro": "^6.3.0",
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
@@ -60,73 +59,73 @@
     "@fortawesome/pro-light-svg-icons": "^6.3.0",
     "@fortawesome/pro-regular-svg-icons": "^6.3.0",
     "@fortawesome/pro-solid-svg-icons": "^6.3.0",
-    "@ngrx/component-store": "^15.3.0",
-    "@ngrx/effects": "^15.3.0",
-    "@ngrx/entity": "^15.3.0",
-    "@ngrx/router-store": "^15.3.0",
-    "@ngrx/store": "^15.3.0",
-    "@ngrx/store-devtools": "^15.3.0",
-    "@nrwl/angular": "15.7.2",
+    "@ngrx/component-store": "^15.4.0",
+    "@ngrx/effects": "^15.4.0",
+    "@ngrx/entity": "^15.4.0",
+    "@ngrx/router-store": "^15.4.0",
+    "@ngrx/store": "^15.4.0",
+    "@ngrx/store-devtools": "^15.4.0",
+    "@nrwl/angular": "~15.8.7",
     "@swimlane/ngx-charts": "^20.1.2",
     "axios": "^0.19.2",
     "core-js": "^2.5.4",
     "firebase": "^9.17.1",
     "firebase-admin": "^11.5.0",
+    "firebase-functions": "^4.2.1",
+    "nodemailer": "^6.9.1",
     "luxon": "^3.2.1",
-    "mysql": "^2.18.1",
     "ngx-infinite-scroll": "^15.0.0",
-    "rxjs": "~7.5.0",
-    "tslib": "^2.0.0",
-    "zone.js": "0.11.4"
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1501.6",
-    "@angular-devkit/build-angular": "15.1.6",
-    "@angular-devkit/core": "~15.1.0",
-    "@angular-eslint/eslint-plugin": "15.0.0",
-    "@angular-eslint/eslint-plugin-template": "15.0.0",
-    "@angular-eslint/template-parser": "15.0.0",
-    "@angular/cli": "~15.1.0",
-    "@angular/compiler-cli": "15.1.5",
-    "@angular/language-service": "15.1.5",
+    "@simondotm/nx-firebase": "^1.0.0",
+    "@angular-devkit/build-angular": "~15.2.4",
+    "@angular-devkit/core": "~15.2.0",
+    "@angular-devkit/schematics": "~15.2.0",
+    "@angular-eslint/eslint-plugin": "~15.2.1",
+    "@angular-eslint/eslint-plugin-template": "~15.2.1",
+    "@angular-eslint/template-parser": "~15.2.1",
+    "@angular/cli": "~15.2.0",
+    "@angular/compiler-cli": "~15.2.0",
+    "@angular/language-service": "~15.2.0",
     "@ngrx/schematics": "^15.3.0",
     "@ngrx/store-devtools": "^15.3.0",
-    "@nrwl/cli": "15.7.2",
-    "@nrwl/cypress": "15.7.2",
-    "@nrwl/eslint-plugin-nx": "15.7.2",
-    "@nrwl/jest": "15.7.2",
-    "@nrwl/linter": "15.7.2",
-    "@nrwl/node": "15.7.2",
-    "@nrwl/workspace": "15.7.2",
-    "@schematics/angular": "~15.1.0",
-    "@types/jest": "28.1.8",
+    "@nrwl/cypress": "15.8.6",
+    "@nrwl/devkit": "15.8.6",
+    "@nrwl/esbuild": "15.8.6",
+    "@nrwl/eslint-plugin-nx": "15.8.6",
+    "@nrwl/jest": "15.8.6",
+    "@nrwl/js": "15.8.6",
+    "@nrwl/linter": "15.8.6",
+    "@nrwl/node": "15.8.6",
+    "@nrwl/workspace": "15.8.6",
+    "@schematics/angular": "~15.2.4",
+    "@types/jest": "^29.5.0",
     "@types/luxon": "^3.2.0",
-    "@types/mysql": "^2.15.15",
     "@types/node": "18.7.1",
     "@typescript-eslint/eslint-plugin": "5.52.0",
     "@typescript-eslint/parser": "5.52.0",
-    "autoprefixer": "^10.4.13",
-    "codelyzer": "6.0.1",
     "cypress": "^3.8.2",
     "dotenv": "10.0.0",
     "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.11.2",
+    "firebase-functions-test": "^0.2.0",
     "firebase-tools": "^11.23.1",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.2.0",
-    "jest": "28.1.3",
-    "jest-environment-jsdom": "28.1.3",
-    "jest-preset-angular": "12.2.3",
-    "nx": "15.7.2",
+    "jest": "^29.4.1",
+    "jest-environment-jsdom": "^29.4.1",
+    "jest-environment-node": "^29.4.1",
+    "jest-preset-angular": "~13.0.0",
+    "nx": "15.8.6",
     "postcss": "^8.4.21",
     "prettier": "2.8.4",
     "tailwindcss": "^3.2.7",
-    "ts-jest": "28.0.8",
+    "kill-port": "^2.0.1",
+    "ts-jest": "^29.0.5",
     "ts-node": "10.9.1",
-    "tslint": "6.1.3",
-    "typescript": "4.8.4",
+    "typescript": "~4.9.5",
     "webpack-bundle-analyzer": "^4.8.0"
   }
 }
