@@ -3,7 +3,7 @@ import { getQualifyResults, getRaceResults } from '@f2020/ergast';
 import { firestoreUtils } from './converter/firestore-utils';
 import { firebaseApp } from './firebase';
 
-export const buildPreviousRaceReult = async (seasonId: number) => {
+export const buildPreviousRaceResult = async (seasonId: number) => {
 
   const results: IRaceResult[] = await (await getRaceResults(seasonId.toString())).map(mapper.raceResult);
   const qualification: IQualifyResult[] = await (await getQualifyResults(seasonId.toString())).map(mapper.qualifyResult);
@@ -16,4 +16,4 @@ export const buildPreviousRaceReult = async (seasonId: number) => {
     return firebaseApp.datebase.doc(`seasons/${seasonId + 1}/lastYear/${result.countryCode}`).set(firestoreUtils.convertDateTimes(concat));
   }));
 
-}
+};
