@@ -1,24 +1,30 @@
+import { buildDrivers } from "./app/drivers";
+import { buildPreviousRaceResult } from "./app/previous-year-race-result";
+import { buildNewSeason } from "./app/season";
+import { assignTeamsToSeason, getTeams, writeTeams } from "./app/teams";
 import { environment } from "./environments/environment";
 
 
 
 
 const seasonId = parseInt(environment.season);
-// getTeams(seasonId - 1)
-//   .then(teams => writeTeams(seasonId, teams))
-//   .then(count => console.log(`Wrote ${count} teams`));
+console.log(`Building season ${seasonId}`);
+
+buildDrivers()
+  .then(count => console.log(`Wrote ${count} drivers`))
+  .catch(error => console.error('Completed with errors', error));
+getTeams(seasonId - 1)
+  .then(teams => writeTeams(seasonId, teams))
+  .then(count => console.log(`Wrote ${count} teams`));
 
 
 // import { environment } from "./environments/environment";
 
-// buildNewSeason(environment.season)
-//   .then(() => assignTeamsToSeason(seasonId))
-//   .then(_ => console.log('Completed', _))
-//   .catch(error => console.error('Completed with errors', error));
+buildNewSeason(environment.season)
+  .then(() => assignTeamsToSeason(seasonId))
+  .then(_ => console.log('Completed', _))
+  .catch(error => console.error('Completed with errors', error));
 
-// buildDrivers()
-//   .then(count => console.log(`Wrote ${count} drivers`))
-//   .catch(error => console.error('Completed with errors', error));
 // Bare skriv de først ti ud
 // readUser().then(transactions => console.log(transactions.slice(0, 10)));
 
@@ -34,4 +40,4 @@ const seasonId = parseInt(environment.season);
 // appendRaces('2021').then(_ => console.log('Added races', _));
 //  assignTeamsToSeason(2023);
 
-// buildPreviousRaceReult(seasonId - 1).then(() => console.log(`Build previous season`));
+buildPreviousRaceResult(seasonId - 1).then(() => console.log(`Build previous season`));
