@@ -1,258 +1,258 @@
-export const countries = {
-	"Afghanistan": "AF",
-	"Aland Islands": "AX",
-	"Albania": "AL",
-	"Algeria": "DZ",
-	"American Samoa": "AS",
-	"Andorra": "AD",
-	"Angola": "AO",
-	"Anguilla": "AI",
-	"Antarctica": "AQ",
-	"Antigua and Barbuda": "AG",
-	"Argentina": "AR",
-	"Armenia": "AM",
-	"Aruba": "AW",
-	"Australia": "AU",
-	"Austria": "AT",
-	"Azerbaijan": "AZ",
-	"Bahamas": "BS",
-	"Bahrain": "BH",
-	"Bangladesh": "BD",
-	"Barbados": "BB",
-	"Belarus": "BY",
-	"Belgium": "BE",
-	"Belize": "BZ",
-	"Benin": "BJ",
-	"Bermuda": "BM",
-	"Bhutan": "BT",
-	"Bolivia": "BO",
-	"Bonaire, Saint Eustatius and Saba ": "BQ",
-	"Bosnia and Herzegovina": "BA",
-	"Botswana": "BW",
-	"Bouvet Island": "BV",
-	"Brazil": "BR",
-	"British Indian Ocean Territory": "IO",
-	"British Virgin Islands": "VG",
-	"Brunei": "BN",
-	"Bulgaria": "BG",
-	"Burkina Faso": "BF",
-	"Burundi": "BI",
-	"Cambodia": "KH",
-	"Cameroon": "CM",
-	"Canada": "CA",
-	"Cape Verde": "CV",
-	"Cayman Islands": "KY",
-	"Central African Republic": "CF",
-	"Chad": "TD",
-	"Chile": "CL",
-	"China": "CN",
-	"Christmas Island": "CX",
-	"Cocos Islands": "CC",
-	"Colombia": "CO",
-	"Comoros": "KM",
-	"Cook Islands": "CK",
-	"Costa Rica": "CR",
-	"Croatia": "HR",
-	"Cuba": "CU",
-	"Curacao": "CW",
-	"Cyprus": "CY",
-	"Czech Republic": "CZ",
-	"Democratic Republic of the Congo": "CD",
-	"Denmark": "DK",
-	"Djibouti": "DJ",
-	"Dominica": "DM",
-	"Dominican Republic": "DO",
-	"East Timor": "TL",
-	"Ecuador": "EC",
-	"Egypt": "EG",
-	"El Salvador": "SV",
-	"Equatorial Guinea": "GQ",
-	"Eritrea": "ER",
-	"Estonia": "EE",
-	"Ethiopia": "ET",
-	"Falkland Islands": "FK",
-	"Faroe Islands": "FO",
-	"Fiji": "FJ",
-	"Finland": "FI",
-	"France": "FR",
-	"French Guiana": "GF",
-	"French Polynesia": "PF",
-	"French Southern Territories": "TF",
-	"Gabon": "GA",
-	"Gambia": "GM",
-	"Georgia": "GE",
-	"Germany": "DE",
-	"Ghana": "GH",
-	"Gibraltar": "GI",
-	"Greece": "GR",
-	"Greenland": "GL",
-	"Grenada": "GD",
-	"Guadeloupe": "GP",
-	"Guam": "GU",
-	"Guatemala": "GT",
-	"Guernsey": "GG",
-	"Guinea": "GN",
-	"Guinea-Bissau": "GW",
-	"Guyana": "GY",
-	"Haiti": "HT",
-	"Heard Island and McDonald Islands": "HM",
-	"Honduras": "HN",
-	"Hong Kong": "HK",
-	"Hungary": "HU",
-	"Iceland": "IS",
-	"India": "IN",
-	"Indonesia": "ID",
-	"Iran": "IR",
-	"Iraq": "IQ",
-	"Ireland": "IE",
-	"Isle of Man": "IM",
-	"Israel": "IL",
-	"Italy": "IT",
-	"Ivory Coast": "CI",
-	"Jamaica": "JM",
-	"Japan": "JP",
-	"Jersey": "JE",
-	"Jordan": "JO",
-	"Kazakhstan": "KZ",
-	"Kenya": "KE",
-	"Kiribati": "KI",
-	"Kosovo": "XK",
-	"Kuwait": "KW",
-	"Kyrgyzstan": "KG",
-	"Laos": "LA",
-	"Latvia": "LV",
-	"Lebanon": "LB",
-	"Lesotho": "LS",
-	"Liberia": "LR",
-	"Libya": "LY",
-	"Liechtenstein": "LI",
-	"Lithuania": "LT",
-	"Luxembourg": "LU",
-	"Macao": "MO",
-	"Macedonia": "MK",
-	"Madagascar": "MG",
-	"Malawi": "MW",
-	"Malaysia": "MY",
-	"Maldives": "MV",
-	"Mali": "ML",
-	"Malta": "MT",
-	"Marshall Islands": "MH",
-	"Martinique": "MQ",
-	"Mauritania": "MR",
-	"Mauritius": "MU",
-	"Mayotte": "YT",
-	"Mexico": "MX",
-	"Micronesia": "FM",
-	"Moldova": "MD",
-	"Monaco": "MC",
-	"Mongolia": "MN",
-	"Montenegro": "ME",
-	"Montserrat": "MS",
-	"Morocco": "MA",
-	"Mozambique": "MZ",
-	"Myanmar": "MM",
-	"Namibia": "NA",
-	"Nauru": "NR",
-	"Nepal": "NP",
-	"Netherlands": "NL",
-	"New Caledonia": "NC",
-	"New Zealand": "NZ",
-	"Nicaragua": "NI",
-	"Niger": "NE",
-	"Nigeria": "NG",
-	"Niue": "NU",
-	"Norfolk Island": "NF",
-	"North Korea": "KP",
-	"Northern Mariana Islands": "MP",
-	"Norway": "NO",
-	"Oman": "OM",
-	"Pakistan": "PK",
-	"Palau": "PW",
-	"Palestinian Territory": "PS",
-	"Panama": "PA",
-	"Papua New Guinea": "PG",
-	"Paraguay": "PY",
-	"Peru": "PE",
-	"Philippines": "PH",
-	"Pitcairn": "PN",
-	"Poland": "PL",
-	"Portugal": "PT",
-	"Puerto Rico": "PR",
-	"Qatar": "QA",
-	"Republic of the Congo": "CG",
-	"Reunion": "RE",
-	"Romania": "RO",
-	"Russia": "RU",
-	"Rwanda": "RW",
-	"Saint Barthelemy": "BL",
-	"Saint Helena": "SH",
-	"Saint Kitts and Nevis": "KN",
-	"Saint Lucia": "LC",
-	"Saint Martin": "MF",
-	"Saint Pierre and Miquelon": "PM",
-	"Saint Vincent and the Grenadines": "VC",
-	"Samoa": "WS",
-	"San Marino": "SM",
-	"Sao Tome and Principe": "ST",
-	"Saudi Arabia": "SA",
-	"Senegal": "SN",
-	"Serbia": "RS",
-	"Seychelles": "SC",
-	"Sierra Leone": "SL",
-	"Singapore": "SG",
-	"Sint Maarten": "SX",
-	"Slovakia": "SK",
-	"Slovenia": "SI",
-	"Solomon Islands": "SB",
-	"Somalia": "SO",
-	"South Africa": "ZA",
-	"South Georgia and the South Sandwich Islands": "GS",
-	"South Korea": "KR",
-	"South Sudan": "SS",
-	"Spain": "ES",
-	"Sri Lanka": "LK",
-	"Sudan": "SD",
-	"Suriname": "SR",
-	"Svalbard and Jan Mayen": "SJ",
-	"Swaziland": "SZ",
-	"Sweden": "SE",
-	"Switzerland": "CH",
-	"Syria": "SY",
-	"Taiwan": "TW",
-	"Tajikistan": "TJ",
-	"Tanzania": "TZ",
-	"Thailand": "TH",
-	"Togo": "TG",
-	"Tokelau": "TK",
-	"Tonga": "TO",
-	"Trinidad and Tobago": "TT",
-	"Tunisia": "TN",
-	"Turkey": "TR",
-	"Turkmenistan": "TM",
-	"Turks and Caicos Islands": "TC",
-	"Tuvalu": "TV",
-	"U.S. Virgin Islands": "VI",
-	"Uganda": "UG",
-	"Ukraine": "UA",
-	"United Arab Emirates": "AE",
-	"United Kingdom": "GB",
-	"UK": "GB",
-	"United States Minor Outlying Islands": "UM",
-	"USA": "US",
-	"UAE": "AE",
-	"United States": "US",
-	"Uruguay": "UY",
-	"Uzbekistan": "UZ",
-	"Vanuatu": "VU",
-	"Vatican": "VA",
-	"Venezuela": "VE",
-	"Vietnam": "VN",
-	"Wallis and Futuna": "WF",
-	"Western Sahara": "EH",
-	"Yemen": "YE",
-	"Zambia": "ZM",
-	"Zimbabwe": "ZW"
-}
+export const countries: Record<string, string> = {
+  "Afghanistan": "AF",
+  "Aland Islands": "AX",
+  "Albania": "AL",
+  "Algeria": "DZ",
+  "American Samoa": "AS",
+  "Andorra": "AD",
+  "Angola": "AO",
+  "Anguilla": "AI",
+  "Antarctica": "AQ",
+  "Antigua and Barbuda": "AG",
+  "Argentina": "AR",
+  "Armenia": "AM",
+  "Aruba": "AW",
+  "Australia": "AU",
+  "Austria": "AT",
+  "Azerbaijan": "AZ",
+  "Bahamas": "BS",
+  "Bahrain": "BH",
+  "Bangladesh": "BD",
+  "Barbados": "BB",
+  "Belarus": "BY",
+  "Belgium": "BE",
+  "Belize": "BZ",
+  "Benin": "BJ",
+  "Bermuda": "BM",
+  "Bhutan": "BT",
+  "Bolivia": "BO",
+  "Bonaire, Saint Eustatius and Saba ": "BQ",
+  "Bosnia and Herzegovina": "BA",
+  "Botswana": "BW",
+  "Bouvet Island": "BV",
+  "Brazil": "BR",
+  "British Indian Ocean Territory": "IO",
+  "British Virgin Islands": "VG",
+  "Brunei": "BN",
+  "Bulgaria": "BG",
+  "Burkina Faso": "BF",
+  "Burundi": "BI",
+  "Cambodia": "KH",
+  "Cameroon": "CM",
+  "Canada": "CA",
+  "Cape Verde": "CV",
+  "Cayman Islands": "KY",
+  "Central African Republic": "CF",
+  "Chad": "TD",
+  "Chile": "CL",
+  "China": "CN",
+  "Christmas Island": "CX",
+  "Cocos Islands": "CC",
+  "Colombia": "CO",
+  "Comoros": "KM",
+  "Cook Islands": "CK",
+  "Costa Rica": "CR",
+  "Croatia": "HR",
+  "Cuba": "CU",
+  "Curacao": "CW",
+  "Cyprus": "CY",
+  "Czech Republic": "CZ",
+  "Democratic Republic of the Congo": "CD",
+  "Denmark": "DK",
+  "Djibouti": "DJ",
+  "Dominica": "DM",
+  "Dominican Republic": "DO",
+  "East Timor": "TL",
+  "Ecuador": "EC",
+  "Egypt": "EG",
+  "El Salvador": "SV",
+  "Equatorial Guinea": "GQ",
+  "Eritrea": "ER",
+  "Estonia": "EE",
+  "Ethiopia": "ET",
+  "Falkland Islands": "FK",
+  "Faroe Islands": "FO",
+  "Fiji": "FJ",
+  "Finland": "FI",
+  "France": "FR",
+  "French Guiana": "GF",
+  "French Polynesia": "PF",
+  "French Southern Territories": "TF",
+  "Gabon": "GA",
+  "Gambia": "GM",
+  "Georgia": "GE",
+  "Germany": "DE",
+  "Ghana": "GH",
+  "Gibraltar": "GI",
+  "Greece": "GR",
+  "Greenland": "GL",
+  "Grenada": "GD",
+  "Guadeloupe": "GP",
+  "Guam": "GU",
+  "Guatemala": "GT",
+  "Guernsey": "GG",
+  "Guinea": "GN",
+  "Guinea-Bissau": "GW",
+  "Guyana": "GY",
+  "Haiti": "HT",
+  "Heard Island and McDonald Islands": "HM",
+  "Honduras": "HN",
+  "Hong Kong": "HK",
+  "Hungary": "HU",
+  "Iceland": "IS",
+  "India": "IN",
+  "Indonesia": "ID",
+  "Iran": "IR",
+  "Iraq": "IQ",
+  "Ireland": "IE",
+  "Isle of Man": "IM",
+  "Israel": "IL",
+  "Italy": "IT",
+  "Ivory Coast": "CI",
+  "Jamaica": "JM",
+  "Japan": "JP",
+  "Jersey": "JE",
+  "Jordan": "JO",
+  "Kazakhstan": "KZ",
+  "Kenya": "KE",
+  "Kiribati": "KI",
+  "Kosovo": "XK",
+  "Kuwait": "KW",
+  "Kyrgyzstan": "KG",
+  "Laos": "LA",
+  "Latvia": "LV",
+  "Lebanon": "LB",
+  "Lesotho": "LS",
+  "Liberia": "LR",
+  "Libya": "LY",
+  "Liechtenstein": "LI",
+  "Lithuania": "LT",
+  "Luxembourg": "LU",
+  "Macao": "MO",
+  "Macedonia": "MK",
+  "Madagascar": "MG",
+  "Malawi": "MW",
+  "Malaysia": "MY",
+  "Maldives": "MV",
+  "Mali": "ML",
+  "Malta": "MT",
+  "Marshall Islands": "MH",
+  "Martinique": "MQ",
+  "Mauritania": "MR",
+  "Mauritius": "MU",
+  "Mayotte": "YT",
+  "Mexico": "MX",
+  "Micronesia": "FM",
+  "Moldova": "MD",
+  "Monaco": "MC",
+  "Mongolia": "MN",
+  "Montenegro": "ME",
+  "Montserrat": "MS",
+  "Morocco": "MA",
+  "Mozambique": "MZ",
+  "Myanmar": "MM",
+  "Namibia": "NA",
+  "Nauru": "NR",
+  "Nepal": "NP",
+  "Netherlands": "NL",
+  "New Caledonia": "NC",
+  "New Zealand": "NZ",
+  "Nicaragua": "NI",
+  "Niger": "NE",
+  "Nigeria": "NG",
+  "Niue": "NU",
+  "Norfolk Island": "NF",
+  "North Korea": "KP",
+  "Northern Mariana Islands": "MP",
+  "Norway": "NO",
+  "Oman": "OM",
+  "Pakistan": "PK",
+  "Palau": "PW",
+  "Palestinian Territory": "PS",
+  "Panama": "PA",
+  "Papua New Guinea": "PG",
+  "Paraguay": "PY",
+  "Peru": "PE",
+  "Philippines": "PH",
+  "Pitcairn": "PN",
+  "Poland": "PL",
+  "Portugal": "PT",
+  "Puerto Rico": "PR",
+  "Qatar": "QA",
+  "Republic of the Congo": "CG",
+  "Reunion": "RE",
+  "Romania": "RO",
+  "Russia": "RU",
+  "Rwanda": "RW",
+  "Saint Barthelemy": "BL",
+  "Saint Helena": "SH",
+  "Saint Kitts and Nevis": "KN",
+  "Saint Lucia": "LC",
+  "Saint Martin": "MF",
+  "Saint Pierre and Miquelon": "PM",
+  "Saint Vincent and the Grenadines": "VC",
+  "Samoa": "WS",
+  "San Marino": "SM",
+  "Sao Tome and Principe": "ST",
+  "Saudi Arabia": "SA",
+  "Senegal": "SN",
+  "Serbia": "RS",
+  "Seychelles": "SC",
+  "Sierra Leone": "SL",
+  "Singapore": "SG",
+  "Sint Maarten": "SX",
+  "Slovakia": "SK",
+  "Slovenia": "SI",
+  "Solomon Islands": "SB",
+  "Somalia": "SO",
+  "South Africa": "ZA",
+  "South Georgia and the South Sandwich Islands": "GS",
+  "South Korea": "KR",
+  "South Sudan": "SS",
+  "Spain": "ES",
+  "Sri Lanka": "LK",
+  "Sudan": "SD",
+  "Suriname": "SR",
+  "Svalbard and Jan Mayen": "SJ",
+  "Swaziland": "SZ",
+  "Sweden": "SE",
+  "Switzerland": "CH",
+  "Syria": "SY",
+  "Taiwan": "TW",
+  "Tajikistan": "TJ",
+  "Tanzania": "TZ",
+  "Thailand": "TH",
+  "Togo": "TG",
+  "Tokelau": "TK",
+  "Tonga": "TO",
+  "Trinidad and Tobago": "TT",
+  "Tunisia": "TN",
+  "Turkey": "TR",
+  "Turkmenistan": "TM",
+  "Turks and Caicos Islands": "TC",
+  "Tuvalu": "TV",
+  "U.S. Virgin Islands": "VI",
+  "Uganda": "UG",
+  "Ukraine": "UA",
+  "United Arab Emirates": "AE",
+  "United Kingdom": "GB",
+  "UK": "GB",
+  "United States Minor Outlying Islands": "UM",
+  "USA": "US",
+  "UAE": "AE",
+  "United States": "US",
+  "Uruguay": "UY",
+  "Uzbekistan": "UZ",
+  "Vanuatu": "VU",
+  "Vatican": "VA",
+  "Venezuela": "VE",
+  "Vietnam": "VN",
+  "Wallis and Futuna": "WF",
+  "Western Sahara": "EH",
+  "Yemen": "YE",
+  "Zambia": "ZM",
+  "Zimbabwe": "ZW"
+} as const;
 
 const countryInfo = [
   {
@@ -1991,7 +1991,7 @@ const countryInfo = [
     "en_short_name": "Zimbabwe",
     "nationality": "Zimbabwean"
   }
-]
+] as const;
 
-export const getByNationality = (nationality: string): string => countryInfo.find(ci => ci.nationality === nationality)?.alpha_2_code;
+export const getByNationality = (nationality: string): string | undefined => countryInfo.find(ci => ci.nationality === nationality)?.alpha_2_code;
 
