@@ -5,9 +5,9 @@ export const driverQualifying = (result: ErgastDriverQualifying): IDriverQualify
   return {
     driver: driver(result.Driver),
     position: parseInt(result.position, 10),
-    q1: result.Q1 ? qualifyToMillis(result.Q1) : null,
-    q2: result.Q2 ? qualifyToMillis(result.Q2) : null,
-    q3: result.Q3 ? qualifyToMillis(result.Q3) : null,
+    q1: result.Q1 ? qualifyToMillis(result.Q1) : -1,
+    q2: result.Q2 ? qualifyToMillis(result.Q2) : undefined,
+    q3: result.Q3 ? qualifyToMillis(result.Q3) : undefined,
   };
 };
 
