@@ -2,7 +2,7 @@ export interface ITeam {
   constructorId: string;
   name: string;
   url: string;
-  countryCode: string;
+  countryCode?: string;
   points: number;
   drivers: string[];
 }
